-Fix cast warnings
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@7253 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/os/win32/thd_priority.cpp
@@ -50,7 +50,7 @@ MemoryPool * ThreadPriorityScheduler::pool = 0;
 ThreadPriorityScheduler * ThreadPriorityScheduler::chain = 0;
 ThreadPriorityScheduler * ThreadPriorityScheduler::news = 0;
 BOOLEAN ThreadPriorityScheduler::initialized = FALSE;
-DWORD ThreadPriorityScheduler::specific_key = -1;
+DWORD ThreadPriorityScheduler::specific_key = (DWORD) -1;
 BOOLEAN ThreadPriorityScheduler::shutdown = FALSE;
 
 //____________________________________________________________
